+++ R/filled_contour_wrblr_int.R
@@ -1,6 +1,6 @@
 #internal warbleR function, not to be called by users. It is a modified version of seewave::filled.contour.modif2
 # that allows to plot spectrograms on top of each other. 
-filled.contour.color.INTFUN <- function (x = seq(0, 1, len = nrow(z)), y = seq(0, 1, len = ncol(z)), 
+filled_contour_wrblr_int <- function (x = seq(0, 1, len = nrow(z)), y = seq(0, 1, len = ncol(z)), 
           z, xlim = range(x, finite = TRUE), ylim = range(y, finite = TRUE), col.lab, colaxis,
           zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels), add = FALSE,
           nlevels = 20, color.palette = cm.colors, col = color.palette(length(levels) - 
@@ -24,10 +24,10 @@ filled.contour.color.INTFUN <- function (x = seq(0, 1, len = nrow(z)), y = seq(0
   }
   if (any(diff(x) <= 0) || any(diff(y) <= 0)) 
     stop("increasing 'x' and 'y' values expected")
-  if(!add) {plot.new()
+  if(!add) plot.new()
   plot.window(xlim, ylim, "", xaxs = xaxs, yaxs = yaxs, asp = asp)
   if (!is.matrix(z) || nrow(z) <= 1 || ncol(z) <= 1) 
-    stop("no proper 'z' matrix specified")}
+    stop("no proper 'z' matrix specified")
   if (!is.double(z)) 
     storage.mode(z) <- "double"
   .filled.contour(as.double(x), as.double(y), z, as.double(levels),