+++ R/filtersels.R
@@ -31,9 +31,10 @@
 #'  image files should be in the working directory (or the directory provided in 'path').
 #' @export
 #' @name filtersels
-#' @examples{ 
+#' @examples
+#' \dontrun{ 
 #' # First set temporary folder
-#' setwd(tempdir())
+#' # setwd(tempdir())
 #' 
 #' # save wav file examples
 #' data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "selec.table"))