+++ R/snrspecs.R
@@ -71,9 +71,9 @@
 #'   Title font size, inner.mar and outer.mar (from \code{mar} and \code{oma} in \code{par}) don't work well
 #'   when osci or sc = \code{TRUE}, this may take some optimization by the user.
 #' @examples
-#' {
+#' \dontrun{
 #' # Set temporary working directory
-#' setwd(tempdir())
+#' # setwd(tempdir())
 #'  
 #' data(list = c("Phae.long1", "Phae.long2", "selec.table"))
 #' writeWave(Phae.long1, "Phae.long1.wav") #save sound.files